MAINT: refactoring, major cleanup
